Fix search function type mismatch causing 500 errors
CRITICAL BUG FIX: Search API was returning 500 errors due to
PostgreSQL type mismatch in search_help_articles function.

Error:
"structure of query does not match function result type"
"Returned type double precision does not match expected type real"

Root Cause:
- Function declared rank as REAL
- PostgreSQL's GREATEST() returns DOUBLE PRECISION
- Type mismatch caused function to fail

Solution:
Changed rank type from REAL to DOUBLE PRECISION in function signature.

This fixes the 500 errors when searching for help articles.

+++ src/lib/db/migrations/0023_add_category_slug_to_search.sql
@@ -16,7 +16,7 @@ RETURNS TABLE (
   slug VARCHAR,
   category_name VARCHAR,
   category_slug VARCHAR,
-  rank REAL
+  rank DOUBLE PRECISION
 ) AS $$
 BEGIN
   RETURN QUERY
